:hammer: Refactored AUC code
* The code belongs to prROC and ROC files. The goal is to write two classes:

    1. ROC/prROC
    2. AUC for each function
